+++ xbird-cloud-netflix/xbird-cloud-netflix-eureka-client/src/main/java/com/xbird/cloud/netflix/eureka/client/EurekaClientApplication.java
@@ -1,0 +1,12 @@
+package com.xbird.cloud.netflix.eureka.client;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EurekaClientApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(EurekaClientApplication.class, args);
+  }
+}