+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - 初始化。
 - [新增] 添加 Spring Cloud Config 项目。
+- [新增] 添加 Spring Cloud Netflix 项目。